Create the core projects of the solution
The Models project is at the center of the application. It defines the
business objects we manipulate, ie what a task is, what an overlap is,
etc

The Logic project contains the behaviors and rules we apply to the objects.

The Definitions contains only interfaces for IoC pattern and upcoming
tests.

The Api project is the ASP.NET Core part serving our logic as an API.